fix rag eval llm dispatch and workflow reconciliation

+++ src/contexts/knowledge_workbench/rag_eval/infrastructure/postgres/postgres_workbench_rag_eval_repository.py
@@ -651,7 +651,7 @@ class PostgresWorkbenchRagEvalRepository(WorkbenchRagEvalRepositoryPort):
                     COALESCE(
                         SUM(question_count),
                         0
-                    ) AS total_questions,
+                    )::bigint AS total_questions,
                     COUNT(*) AS represented_entry_count,
                     COUNT(*) FILTER (
                         WHERE question_count = $2
